@@ -1,3 +1,5 @@
+/* global View */
+// eslint-disable-next-line no-unused-vars
 class Scene {
   constructor(uid, cfgs) {
     const scene = cfgs.scenes[uid];
@@ -6,8 +8,8 @@ class Scene {
     this.name = scene.name;
     this.className = scene.className;
     this.uid = uid;
-    this.nextUid = uid == sceneLength - 1 ? 0 : uid + 1;
-    this.previousUid = uid == 0 ? sceneLength - 1 : uid - 1;
+    this.nextUid = uid === sceneLength - 1 ? 0 : uid + 1;
+    this.previousUid = uid === 0 ? sceneLength - 1 : uid - 1;
     this.viewNames =
       !scene.views || scene.views.length == 0
         ? cfgs.views.map((v) => v.name)
@@ -17,10 +19,9 @@ class Scene {
 
   draw(events) {
     this.clearViews();
-    const viewId = 0;
     this.viewNames.forEach((viewName) => {
       const viewConfig = this.config.views.find((config) => {
-        if (config.name == viewName) return true;
+        if (config.name === viewName) return true;
       });
       viewConfig.sceneClassName = this.className;
       const view = View.makeByName(viewConfig, events);
