@@ -1,5 +1,5 @@
-/* global moment */
-
+/* global Slot ViewCurrent ViewUpcoming ViewMonth ViewDaily ViewWeekly ViewMonthly ViewWeek */
+// eslint-disable-next-line no-unused-vars
 class View {
   constructor(config, events) {
     this.config = config;
@@ -87,7 +87,7 @@ class View {
     const calendarFilter =
       Array.isArray(calendars) && calendars.length > 0
         ? (e) => calendars.indexOf(e.calendarName) >= 0
-        : (e) => true;
+        : () => true;
     let filtered = events.filter(calendarFilter);
     if (this.config.filterPassedEvent) {
       const now = moment().format("X");
@@ -310,390 +310,6 @@ class View {
     for (let i = 0; i < this.slots.length; i++) {
       const slot = this.slots[i];
       slot.drawEvents();
-    }
-  }
-}
-
-class ViewPeriod extends View {
-  constructor(config, events) {
-    super(config, events);
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("period");
-  }
-
-  viewDomType(viewDom) {
-    viewDom.classList.add(this.config.type);
-  }
-
-  adjustSlotWidth(slotDom, count) {
-    if (this.config.type == "row") slotDom.style.width = `${100 / count - 3}%`;
-  }
-}
-
-class ViewAgenda extends View {
-  constructor(config, events) {
-    super(config, events);
-    this.slotUnit = "day";
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("agenda", "period");
-  }
-
-  getSlotCount() {
-    return 1;
-  }
-
-  getStartDay() {
-    const now = moment().locale(this.locale);
-    return now;
-  }
-
-  getSlotPeriods() {
-    return [
-      {
-        start: moment().locale(this.locale),
-        end: null
-      }
-    ];
-  }
-
-  makeSlotHeader(slot) {
-    const header = slot.headerDom;
-    const title = header.querySelector(".slotTitle");
-    title.innerHTML = this.config.slotTitle;
-  }
-}
-
-class ViewCell extends View {
-  constructor(config, events) {
-    super(config, events);
-    this.slotUnit = "week";
-  }
-
-  makeSlots() {
-    this.contentDom.innerHTML = "";
-    this.slotPeriods = this.getSlotPeriods();
-    this.slots = WeekSlot.factory(this, this.slotPeriods, this.events);
-    for (let i = 0; i < this.slots.length; i++) {
-      const slot = this.slots[i];
-      this.appendSlot(slot);
-    }
-    this.makeModuleTitle();
-  }
-
-  getSlotPeriods() {
-    const showWeekends = this.config.showWeekends;
-    const getSlotPeriod = (tDay, seq) => {
-      const mtd = moment(tDay).locale(this.locale).add(seq, "week");
-      const start = showWeekends
-        ? moment(mtd).startOf("week")
-        : moment(mtd).startOf("week").day(1).startOf("day");
-      const end = showWeekends
-        ? moment(mtd).endOf("week")
-        : moment(mtd).startOf("week").day(5).endOf("day");
-      return {
-        start,
-        end
-      };
-    };
-    const periods = [];
-    const now = moment().locale(this.locale);
-    const targetDay = this.getStartDay();
-    const count = this.getSlotCount();
-    for (let i = 0; i < count; i++) {
-      const period = getSlotPeriod(targetDay, i);
-      periods.push(period);
-    }
-    return periods;
-  }
-
-  getSubSlotPeriods(start) {
-    const showWeekends = this.config.showWeekends;
-    const days = showWeekends ? 7 : 5;
-    const periods = [];
-    const t = start;
-    const startDay = showWeekends
-      ? moment(t).locale(this.locale).startOf("week")
-      : moment(t).locale(this.locale).startOf("week").day(1).startOf("day");
-    for (let i = 0; i < days; i++) {
-      const p = {
-        start: moment(startDay).startOf("day"),
-        end: moment(startDay).endOf("day")
-      };
-      periods.push(p);
-      startDay.add(1, "day");
-    }
-    return periods;
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("weekSlot");
-  }
-
-  viewDomType(viewDom) {
-    viewDom.classList.add("column");
-  }
-
-  adjustSlotHeight(slotDom) {
-    slotDom.style.maxHeight = this.config.slotMaxHeight;
-    slotDom.style.height = this.config.slotMaxHeight;
-  }
-
-  makeCellDomClass(slot, daySeq, weekSeq) {
-    const slotDom = slot.dom;
-    if (daySeq >= 0) slotDom.classList.add(`cellSeq_${daySeq}`);
-    if (weekSeq === 0 && daySeq === 0) {
-      slotDom.classList.add("firstCell");
-    }
-    const day = moment(slot.start).locale(this.locale);
-    const now = moment().locale(this.locale);
-    if (now.format("YYYY") == day.format("YYYY"))
-      slotDom.classList.add("thisyear");
-    if (now.format("M") == day.format("M")) slotDom.classList.add("thismonth");
-    if (now.format("w") == day.format("w")) slotDom.classList.add("thisweek");
-    if (now.format("YYYYMMDD") == day.format("YYYYMMDD"))
-      slotDom.classList.add("today");
-    if (now.format("YYYYMMDD") > day.format("YYYYMMDD"))
-      slotDom.classList.add("passedday");
-    slotDom.classList.add(`weekday_${day.format("E")}`);
-    slotDom.classList.add(`year_${day.format("YYYY")}`);
-    slotDom.classList.add(`month_${day.format("M")}`);
-    slotDom.classList.add(`day_${day.format("D")}`);
-    slotDom.classList.add(`week_${day.format("w")}`);
-    slotDom.classList.add(`dayofyear_${day.format("DDD")}`);
-  }
-
-  makeWeeksMark(start) {
-    const weeks = document.createElement("div");
-    weeks.classList.add("weeksmark");
-    weeks.innerHTML = moment(start)
-      .locale(this.locale)
-      .format(this.config.weeksFormat);
-    return weeks;
-  }
-
-  adjustSlotWidth(slotDom, count) {
-    // if (this.config.type == "row") slotDom.style.width = ((100 / count) - 0.25) + "%"
-  }
-
-  makeSlotHeader(slot) {
-    super.makeSlotHeader(slot);
-    const header = slot.headerDom;
-    const start = slot.start;
-    const altTitle = header.querySelector(".slotAltTitle");
-    if (this.config.slotAltTitle) {
-      altTitle.innerHTML = this.config.slotTitle;
-    } else if (this.config.slotAltTitleFormat) {
-      altTitle.innerHTML = moment(slot.start)
-        .locale(this.locale)
-        .format(this.config.slotAltTitleFormat);
-    }
-  }
-}
-
-class ViewDaily extends ViewPeriod {
-  constructor(config, events) {
-    super(config, events);
-    this.slotUnit = "day";
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("daily");
-
-    const day = moment(slot.start).locale(this.locale);
-    const now = moment().locale(this.locale);
-    if (now.format("YYYY") == day.format("YYYY"))
-      slotDom.classList.add("thisyear");
-    if (now.format("M") == day.format("M")) slotDom.classList.add("thismonth");
-    if (now.format("w") == day.format("w")) slotDom.classList.add("thisweek");
-    if (now.format("YYYYMMDD") == day.format("YYYYMMDD"))
-      slotDom.classList.add("today");
-    if (now.format("YYYYMMDD") > day.format("YYYYMMDD"))
-      slotDom.classList.add("passed");
-    slotDom.classList.add(`weekday_${day.format("E")}`);
-    slotDom.classList.add(`year_${day.format("YYYY")}`);
-    slotDom.classList.add(`month_${day.format("M")}`);
-    slotDom.classList.add(`day_${day.format("D")}`);
-    slotDom.classList.add(`week_${day.format("w")}`);
-    slotDom.classList.add(`dayofyear_${day.format("DDD")}`);
-  }
-}
-
-class ViewWeekly extends ViewPeriod {
-  constructor(config, events) {
-    super(config, events);
-    this.slotUnit = "week";
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("weekly");
-
-    const day = moment(slot.start).locale(this.locale);
-    const dayEnd = moment(slot.end).locale(this.locale);
-    const now = moment().locale(this.locale);
-    if (now.format("YYYY") == day.format("YYYY"))
-      slotDom.classList.add("thisyear");
-    if (now.format("M") == day.format("M")) slotDom.classList.add("thismonth");
-    if (now.format("w") == day.format("w")) slotDom.classList.add("thisweek");
-    if (now.isBetween(day, dayEnd)) slotDom.classList.add("today");
-    slotDom.classList.add(`year_${day.format("YYYY")}`);
-    slotDom.classList.add(`month_${day.format("M")}`);
-    slotDom.classList.add(`week_${day.format("w")}`);
-  }
-
-  makeSlotHeader(slot) {
-    const header = slot.headerDom;
-    const title = header.querySelector(".slotTitle");
-    const subTitle = header.querySelector(".slotSubTitle");
-    if (this.config.slotTitle) {
-      title.innerHTML = this.config.slotTitle;
-    } else if (
-      this.config.slotTitleFormat &&
-      typeof this.config.slotTitleFormat !== "object"
-    ) {
-      title.innerHTML = moment(slot.start)
-        .locale(this.locale)
-        .format(this.config.slotTitleFormat);
-    } else {
-      title.innerHTML = moment(slot.start)
-        .locale(this.locale)
-        .calendar(null, this.config.slotTitleFormat);
-    }
-    if (this.config.slotSubTitle) {
-      subTitle.innerHTML = this.config.slotSubTitle;
-    } else if (
-      this.config.slotSubTitleFormat &&
-      typeof this.config.slotSubTitleFormat !== "object"
-    ) {
-      subTitle.innerHTML = moment(slot.start)
-        .locale(this.locale)
-        .format(this.config.slotSubTitleFormat);
-    } else {
-      subTitle.innerHTML = moment(slot.start)
-        .locale(this.locale)
-        .calendar(null, this.config.slotSubTitleFormat);
-    }
-  }
-}
-
-class ViewMonthly extends ViewPeriod {
-  constructor(config, events) {
-    super(config, events);
-    this.slotUnit = "month";
-  }
-
-  makeSlotDomClass(slot, seq = null) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("monthly");
-    const day = moment(slot.start).locale(this.locale);
-    const dayEnd = moment(slot.end).locale(this.locale);
-    const now = moment().locale(this.locale);
-    if (now.format("YYYY") == day.format("YYYY"))
-      slotDom.classList.add("thisyear");
-    if (now.format("M") == day.format("M")) slotDom.classList.add("thismonth");
-    if (now.isBetween(day, dayEnd)) slotDom.classList.add("today");
-    slotDom.classList.add(`year_${day.format("YYYY")}`);
-    slotDom.classList.add(`month_${day.format("M")}`);
-  }
-}
-
-class ViewCurrent extends ViewAgenda {
-  constructor(config, events) {
-    super(config, events);
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("current");
-  }
-
-  filterEvents(events) {
-    let filtered = super.filterEvents(events);
-    const now = moment();
-    filtered = filtered.filter((e) => {
-      if (now.isBetween(moment.unix(e.startDate), moment.unix(e.endDate)))
-        return true;
-    });
-    return filtered;
-  }
-}
-
-class ViewUpcoming extends ViewAgenda {
-  constructor(config, events) {
-    super(config, events);
-  }
-
-  makeSlotDomClass(slot) {
-    const slotDom = slot.dom;
-    super.makeSlotDomClass(slot);
-    slotDom.classList.add("upcoming");
-  }
-
-  filterEvents(events) {
-    const until = moment().add(this.config.maxDays, "day").endOf("day");
-    let filtered = super.filterEvents(events);
-    filtered = filtered.filter((e) => {
-      if (moment.unix(e.startDate).isBetween(moment(), until)) return true;
-    });
-    return filtered;
-  }
-}
-
-class ViewWeek extends ViewCell {
-  constructor(config, events) {
-    super(config, events);
-  }
-}
-
-class ViewMonth extends ViewCell {
-  constructor(config, events) {
-    super(config, events);
-  }
-
-  getSlotCount() {
-    const startDay = this.getStartDay();
-    const endDay = this.getEndWeek();
-    const diff = endDay.diff(startDay, "week");
-    return diff + 1;
-  }
-
-  getStartDay() {
-    const fromNow = this.config.fromNow;
-    const now = moment().locale(this.locale);
-    return now.add(fromNow, "month").startOf("month").startOf("week");
-  }
-
-  getEndWeek() {
-    const fromNow = this.config.fromNow;
-    const now = moment().locale(this.locale);
-    return now.add(fromNow, "month").endOf("month").startOf("week");
-  }
-
-  makeSlots() {
-    super.makeSlots();
-    if (this.config.monthFormat) {
-      const fromNow = this.config.fromNow;
-      const now = moment().locale(this.locale);
-      now.add(fromNow, "month").startOf("month");
-      const mt = document.createElement("div");
-      mt.innerHTML = now.format(this.config.monthFormat);
-      mt.classList.add("monthViewTitle");
-      this.contentDom.prepend(mt);
     }
   }
 }
